requirements.txt
streamlit
pandas
plotly
tensorflow
scikit-learn
numpy
Pillow
